% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertfigure.R
\name{insertfigure}
\alias{insertfigure}
\alias{EZfigure}
\title{Function to insert a figure and format the caption when using pandoc-crossref.}
\usage{
insertfigure(
  path,
  caption,
  label = NULL,
  shortcaption = NULL,
  unlisted = FALSE,
  ext = "png",
  auto_pdf = getOption("knitr.graphics.auto_pdf", FALSE)
)

EZfigure(caption, shortcaption = NULL)
}
\arguments{
\item{path}{Full path to the image file (with extension), default is the path to the figure of the last chunk}

\item{caption}{A character vector containing the (long) caption of the figure.}

\item{label}{An optional character vector containing the label used for referencing.}

\item{shortcaption}{An optional character vector containing the short caption.}

\item{unlisted}{Either \code{TRUE} or \code{FALSE}, to determine if the figure should be added to the list of figures}

\item{ext}{The file extension, defaults to \code{"png"}}

\item{auto_pdf}{Whether to use PDF images automatically when the output format is LaTeX. If \code{TRUE}, then e.g. \code{foo/bar.png} will be replaced by \code{foo/bar.pdf} if the latter exists. This can be useful since normally PDF images are of higher quality than raster images like PNG, when the output is LaTeX/PDF.}
}
\value{
Command to insert a figure in Markdown
}
\description{
Formats the caption according to the output format.
For the short caption, you need this Lua Filter: https://github.com/pandoc/lua-filters/tree/master/short-captions
The function \code{EZfigure()} is a wrapper function for \code{insertfigure(..., label = opts_current$get("label"), path = knitr::fig_chunk(opts_current$get("label"), ext = "png"), unlisted = FALSE, ext = "png", auto_pdf = TRUE)}. By default, the file path from the previous chunk is used.
}
