% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_after.R
\name{landscape_after}
\alias{landscape_after}
\title{Print the table on an isolated landscape page in PDF
like the function inn the \code{kableExtra} package}
\usage{
landscape_after(kable_input, margin = NULL, clearpage = FALSE)
}
\arguments{
\item{kable_input}{Output of \code{knitr::kable()} with \code{format} specified}

\item{margin}{Customizable page margin for special needs. Values can be
"1cm", "1in" or similar.}

\item{clearpage}{If \code{TRUE} the landscape page is inserted after the page break. Requires the LaTeX package \code{afterpage}.}
}
\description{
This function will put the table on an single landscape page after the next page break.
It's useful for wide tables that can't be printed on a portrait page.
}
\examples{
\dontrun{
landscape_after(knitr::kable(head(mtcars), "latex"), clearpage = TRUE)
}

}
