% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_book.R
\name{latex_book}
\alias{latex_book}
\title{Convert R Markdown to a PDF book LaTeX file}
\usage{
latex_book(
  toc = TRUE,
  number_sections = TRUE,
  fig_caption = TRUE,
  pandoc_args = NULL,
  ...,
  base_format = rmarkdown::pdf_document,
  toc_unnumbered = TRUE,
  toc_appendix = FALSE,
  toc_bib = FALSE,
  quote_footer = NULL,
  highlight_bw = FALSE
)
}
\arguments{
\item{toc}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{number_sections}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{fig_caption}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{pandoc_args}{See
\code{rmarkdown::\link{pdf_document}}, or the documentation of the
\code{base_format} function.}

\item{...}{Other arguments to be passed to \code{base_format}.}

\item{base_format}{An output format function to be used as the base format.}

\item{toc_unnumbered}{Whether to add unnumbered headers to the table of
contents.}

\item{toc_appendix}{Whether to add the appendix to the table of contents.}

\item{toc_bib}{Whether to add the bibliography section to the table of
contents.}

\item{quote_footer}{If a character vector of length 2 and the quote footer
starts with three dashes (\samp{---}), \code{quote_footer[1]} will be
prepended to the footer, and \code{quote_footer[2]} will be appended; if
\code{NULL}, the quote footer will not be processed.}

\item{highlight_bw}{Whether to convert colors for syntax highlighting to
black-and-white (grayscale).}
}
\value{
Command to insert a figure in Markdown
}
\description{
Convert R Markdown files to LaTeX file after resolving the special tokens of \pkg{bookdown} (e.g., the tokens for references and labels) to native LaTeX commands.
Identical to \code{bookdown::pdf_book}, except that the tex file created by Pandoc is not compiled.
}
