% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablecaption.R
\name{tablecaption}
\alias{tablecaption}
\alias{EZtable}
\title{Function to format captions for table when using pandoc-crossref and table-short-captions Lua filter.}
\usage{
tablecaption(caption, label = NULL, shortcaption = NULL, unlisted = FALSE)

EZtable(caption, shortcaption = NULL, label = knitr::opts_current$get("label"))
}
\arguments{
\item{caption}{A character vector containing the (long) caption of the table.}

\item{label}{An optional character vector containing the label used for referencing.}

\item{shortcaption}{An optional character vector containing the short caption.}

\item{unlisted}{Either \code{TRUE} or \code{FALSE}, to determine if the table should be added to the list of tables.}
}
\value{
Formatted table caption
}
\description{
Formats the caption according to the output format.
For the short caption, you need this Lua Filter: https://github.com/pandoc/lua-filters/tree/master/table-short-captions
}
\examples{
tablecaption("Caption of the Table")

}
