% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jr.R
\name{theme_jr}
\alias{theme_jr}
\title{Provides ggplot2 theme theme_jr()}
\usage{
theme_jr(
  base_size = 10,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family (LM Sans 10)}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\value{
Theme for ggplot2
}
\description{
Provides ggplot2 theme theme_jr()
}
\examples{
df <- data.frame(x = runif(10), y = rnorm(10))
require(ggplot2)
ggplot(df, aes(x = x, y = y)) + geom_point() + theme_jr()
}
