% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-info.R
\name{Session_Info}
\alias{Session_Info}
\title{Print session information}
\usage{
Session_Info(
  pkgs = NULL,
  include_base = FALSE,
  include_libs = FALSE,
  check_versions = TRUE
)
}
\arguments{
\item{pkgs}{Either a vector of package names or NULL. If \code{NULL},
displays all loaded packages. If a character vector, also, includes
all dependencies of the package.}

\item{include_base}{Include base packages in summary? By default this is
false since base packages should always match the R version.}

\item{include_libs}{Include the library path of each package in the output. Disabled by default}

\item{check_versions}{Compares the used version of Pandoc with the Pandoc version used to build pandoc-crossref. For more information, see Details}
}
\description{
This is \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}} rewritten to export the version of
Pandoc, pandoc-citeproc and \href{https://github.com/lierdakil/pandoc-crossref}{pandoc-crossref}.
By default it also hides the library paths. Apart from that, it's identical to the original function.
}
\details{
Columns in the \emph{printed} package list:
\itemize{
\item \code{package}: package name
\item \code{*}: whether the package is attached to the search path
\item \code{version}: package version. If the version is marked with \verb{(!)} that
means that the loaded and the on-disk version of the package are
different.
\item \code{date}: when the package was built, if this information is available.
This is the \code{Date/Publication} or the \code{Built} field from
\code{DESCRIPTION}. (These are usually added automatically by R.)
Sometimes this data is not available, then it is \code{NA}.
\item \code{source}: where the package was built or installed from, if available.
Examples: \verb{CRAN (R 3.3.2)}, \verb{Github (r-lib/pkgbuild@8aab60b)},
\code{Bioconductor}, \code{local}.
}

If the Pandoc version used to build the document is different from the version which was used
to compile pandoc-crossref, this is noted in the output if \code{check_versions = TRUE}. This may be helpful
because mixing versions is not supported and "strange things may (and likely will) happen silently"
due to internal changes of Pandoc between versions.

See \code{\link[sessioninfo:package_info]{sessioninfo::package_info()}} for the list of columns in the data frame that
is \emph{returned} (as opposed to \emph{printed}).
}
\examples{
Session_Info()
Session_Info("jRmisc")
}
